Create day.js
Your program will accept a month, day, and year and must produce the day of the week.
For example

What is the month (1 to 12)? 1
What is the day (1 to 31)? 30
What is the year (e.g., 2007) 2007
Total Days Past 766675
Your date falls on a Wednesday


Your program must ensure to account for incorrect input. Specifically, it must
gracefully handle situations where the input is out of valid ranges.
For incorrect values, you should print a message similar to the following:

What is the month (1 to 12)? 123
The month 123 is invalid. Try again
What is the month (1 to 12)? 1
What is the day (1 to 31)? 30
What is the year (e.g., 2007) 2007
Total Days Past 766675
Your date falls on a Wednesday


In addition, you must take into account leap years. The rules of leap years
were developed by the Romans where the following conditions need to be met:

Every 4 years is a leap year
Every 100 years a leap year is skipped unless the year is divisible by 400